bonus14.py refactored functions into separate python files and imported them into bonus14.py
