+++ bonus/bonus14.py
@@ -1,17 +1,7 @@
+from bonus.converters14 import convert
+from bonus.parsers14 import parse
+
 feet_inches = input("Enter feet and inches: ")
-
-
-def parse(feet_inches):
-    parts = feet_inches.split(" ")
-    feet = float(parts[0])
-    inches = float(parts[1])
-    return {"feet":feet, "inches": inches}
-
-
-def convert(feet, inches):
-    meters = feet * 0.3048 + inches * 0.0254
-    return meters
-
 
 parsed = parse(feet_inches)
