+++ bonus/converters14.py
@@ -1,0 +1,3 @@
+def convert(feet, inches):
+    meters = feet * 0.3048 + inches * 0.0254
+    return meters
