+++ bonus/parsers14.py
@@ -1,0 +1,5 @@
+def parse(feet_inches):
+    parts = feet_inches.split(" ")
+    feet = float(parts[0])
+    inches = float(parts[1])
+    return {"feet":feet, "inches": inches}
